[Xamarin.Android.Build.Tasks] Force xbuild to behave like msbuild.
Commit 8643ded9 tried to make sure xbuild worked like msbuild
when it came to `'$(DebugType)' == 'None'`. The "fix" however
was not correct.

When DebugType is None under msbuild, it changes to `portable`
and DebugSymbols get set to 'true'. Also when DebugType is empty
it gets set to `portable` but the DebugSymbols is left alone.

So we need to get xbuild to do the same. Otherwise we get the
wrong runtimes. But by the time we get the the .Common.targets
it is too late to change the `DebugSymbols` value. xbuild does
not take the change. So we need to do this before xbuild gets
a chance to set the value of `DebugSymbols`.

The trick is to alter the properties in the CSharp and Fsharp
targets BEFORE we import the Microsoft targets.