[Xamarin.Android.Build.Tasks] Fix VS DesignTime Build Hang (#985)
FixesL https://bugzilla.xamarin.com/show_bug.cgi?id=60080

In a previous attempt to fix an intellisense build (1cd582ec) we
caused this issue. The DesignTime build does NOT like our AsyncTask
as it locks the IDE.

This commit puts the old code back in place which skips the task
if we are in DesignTime mode. But so we can easily test it, we
emit an message that our test can pick up to ensure we are doing
the correct thing. Without the message we have no way of knowing
if the task is skipped.