[Xamarin.Android.Build.Tasks] Move DesignTime Cache files to designtime folder. (#986)
Fixes: https://devdiv.visualstudio.com/DevDiv/XamarinVS/_queries?id=512752%2F&triage=true&_a=edit

The DesignTime build is generally run before any nuget packages
have been restored. As a result we end up with empty cache files.
Because those files exist and are considered uptodate, when the
main build finally happens they are not re-generated.
This causes errors like

	error: Error: No resource found that matches the given name: attr 'colorAccent'.

on a new default template.

The DesignTime build should have its own set of cache files which
do not conflict with the main build. This way when the main build
runs the cache files will not be in place.