[Xamarin.Android.Build.Tests] Fix BuildAppCheckDebugSymbols for monodroid
The BuildAppCheckDebugSymbols test does not take into account
that when building in debug mode on monodroid the assemblies
are not embedded in the apk. As a result when trying to get
`Library1.pdb` from the apk we get a null back. This causes
a Null Ref exception when we try to get a length.

The fix is to look in the `assets` directory if we can't get
the file from the zip. This will allow the test to work in
both xamarin-android and monodroid.