[xabuild.exe] handle AndroidSdkDirectory like the script
Context:
https://github.com/xamarin/xamarin-android/blob/master/tools/scripts/xabuild#L105

`xabuild`, the shell script, has some logic to set MSBuild properties
for `AndroidSdkDirectory` and `AndroidNdkDirectory`. The issue here is
on Windows, `xabuild.exe` is not invoked from this shell script. So we
need to replicate the logic in C# in `xabuild.exe`, but not break what
was fixed in #977.

Changes:
- check for `ANDROID_SDK_PATH` or `ANDROID_NDK_PATH` environment
variables
- Run `Paths.targets` via a new MSBuild process, the same way as
`xabuild`, the script
- Set the `AndroidSdkDirectory` and `AndroidNdkDirectory` values in the
config file if they are not blank
- `AndroidSdkDirectory` and `AndroidNdkDirectory` can still be
overridden by a user, because incoming values supersede what is found
in the config file