[build] update SDK tools URLs
On Windows, the `lint` tool installed by the build was broken.
This causes several tests to fail on Windows verifying features
around `lint` After looking into it, it appears the URLs used to
download the SDK tools were not found in:

https://dl.google.com/android/repository/repository-12.xml

The correct URLs appear to be of the form `tools_rx.x.x-platform.zip`.
Note Google uses `macosx` in this file name instead of `darwin`.
Not sure what version of the SDK tools were being used previously,
they must be a prerelease or nightly release.

Other changes:
- The command `sdkmanager --license` is now `sdkmanager --update`