[Xamarin.Android.Build.Tasks] Rework AsyncTask to use ConfigureAwait
We should be using `ConfigureAwait(false)` for our `Task.Run`
calls to ensure that the Continuations do NOT run on the
main thread. This is probably the reason why VS locks up
when we call these methods. The Continuation which calls
`Complete` is trying to run on the UI thread, which will
be locked waiting for the Task to complete.

We should also provide the Cancelation Token and the
TaskScheduler for the `Parallel.ForEach` calls.