[build] fix for Xamarin.Android.Build.Tasks re-build
One issue here is that `obj/Debug/Profile.g.cs` is getting regenerated
on every build, forcing `Xamarin.Android.Build.Tasks.csproj` to rebuild
*every* time. It looks like the `IncrementalClean` target was the culprit,
in that it deleted this file after the build completed.

The fix here is to add to the `FileWrites` `ItemGroup`, but it only
seems to work if `FileWrites` includes the absolute path to the file.
Including `$(MSBuildThisFileDirectory)` in front of
`$(_GenerateProfileClass)` makes sure it is the absolute path.