[tests] workaround broken lint.bat on Windows
Context in #992

The lint.bat tool installed to ~\android-toolchain\sdk\tools\bin
appears to be broken on Windows, so a few tests are failing that use it.

There is a workaround to set the `JAVA_OPTS` environment variable
that gets these tests passing on Windows. Hopefully, this is a temporary
workaround until a new version of the SDK tools comes out that fixes
this issue.

One other thing to note here is that the Android SDK installed by Visual
Studio uses the SDK tools version 25.x, which does not appear to have
this issue on windows.