[Xamarin.Android.Build.Tasks] Remove Unsupported Lint Checks (#975)
If a user provides lint checks which are not supported by
the current lint version it will error.

	Invalid id or ategory "StatckFieldLeak"

This is normally ok. But in our build system we sometimes end up
on a bot which has an older version of lint. And in this case
the test fails, which in reality it should ignore the id we
are trying to use.

So this commit addds some validation to the Disabled and Enabled
checks. If its not supported we will remove the check and issue
a warning.