[Xamarin.ProjectTools] Allow msbuild tests to use system x-a. (#1014)
We want to be able to run the msbuild tests against a system
installed xamarin-android. So we need to fallback to using
`msbuild` or `xbuild` if the required `xabuild` is NOT in place.

So if we do a full build `xabuild` will be in `bin\$(Configuration)\bin`
and the tests will run under the local environment. But if we just build
the tests `xabuild` will not exist since its copied as part of the
XABuild project build process.