[tests] fix Java.Interop test results copy
On Windows, under MSBuild, I noticed a clean build was not copying the
test results from `external/Java.Interop` to the root directory. Two
things were causing this to be overlooked: the `<Copy />` task left the
original files from past test runs, and the `@(_JavaInteropTestResults)`
`ItemGroup` was calculated at the top of the file. I *think* this is
somehow working on xbuild.

To fix this, I moved the location of the `@(_JavaInteropTestResults)`
`ItemGroup` to happen during the target, instead of at the top of the
file. I also added a `<Delete />` task after the `<Copy />` to keep a clean
`external/Java.Interop` directory. It doesn't look like xbuild supports
`<Move />`.