[tests] changes to build test APKs with system Xamarin.Android
`Mono.Android-Tests.csproj` needed to use `$(MSBuildExtensionsPath)` as
part of locating `Xamarin.Android.CSharp.targets`. The other test APK
projects were already doing this.

The other change here is to set `XAIntegratedTests` to False by default,
so that it is not required to be set when building the test APKs with
your system Xamarin.Android installation.