[Xamarin.Android.Build.Tasks] Cache AndroidIncludeDebugSymbols (#1015)
The VS team have asked if there is a way to know which runtime
is being used by a particular build configuration.

Since we calculate the `AndroidIncludeDebugSymbols` when the project
is loaded is makes sense to store that in the `build.props` file.
It should not change very often unless the user constantly
updated the project. This way the VS team can read the `build.props`
file and look for the value.

We could have added a new target, but at the point this is needed
VS can not call MSBuild targets or read properties.