[Mono.Android-Tests] Support system Xamarin.Android builds (#1017)
`Mono.Android-Tests.csproj` needed to use `$(MSBuildExtensionsPath)`
as part of locating `Xamarin.Android.CSharp.targets`. The other test
`.apk` projects were already doing this.

The other change here is to set `$(XAIntegratedTests)` to False by
default, so that it is not required to be set when building the test
`.apks` with your system Xamarin.Android installation.

With this change, all on-device `.apk`-based unit tests should be
buildable against a system Xamarin.Android install, and not just the
in-tree build artifacts.

*Note*: Xamarin.Android 8.0 (d15-5) or later is required in order to
build the `.apk`-based unit tests due to commit e9daf5ea, which
updates `Xamarin.Android.NUnitLite.dll`.
The `Xamarin.Android.Bcl-Tests` project requires the new members.