[tests] option to include/exclude categories from Test APKs
In some QA needs to run tests on devices without a network connection,
so an option to disable tests that rely on the network in Mono.Android-Tests
is needed.

NUnit has the option to include or exclude tests based on the
`CategoryAttribute`. Various tests were decorated with
`[Category("InetAccess")]` that appeared to use the internet. To verify
this works, I ran the tests with my internet disabled. The test run for
Mono.Android-Tests went down from 106 to 84 tests, with no failures.

To exclude a category, on Windows:
    msbuild Xamarin.Android.sln /t:RunApkTests /p:ExcludeCategories=InetAccess

On MacOS/Linux:
    make run-apk-tests EXCLUDECATEGORIES=InetAccess

If you want to specify multiple categories, use `:` (colon) to delimit
each category. This delimiter is used for various values throughout
Xamarin.Android's build because it works well from the command line for both
MSBuild and make.

I also added the option specify `Include` as well, but it isn't particularly
useful for this case.