[class-parse] JavaApiParameterNames use `;` for comments (#213)
Change the `JavaDocletType.JavaApiParameterNamesXml` comment
character from `#` to `;`. This is because the parameter description
file format uses `#ctor` for constructors, which obviously interacts
very badly when "`#` until end of line" is a comment.

(Constructors can't be expressed!)
  
This issue was uncovered at API Level 26 which started to contain
generic constructors.