[Tests] Run runtime test in Release/Aot configuration again
Recently the migration of running the apk tests to msbuild
(https://github.com/xamarin/xamarin-android/commit/cdf3bcc11aaf036b68daaec504f6ffa4dbbc120a)
stopped running runtime test with Aot and thus we are missing
performance measurements for Release/Aot.

The reason of it is that `_AotName` property is not set, when calling
MSBuild task to run the apk tests. This patch sets the `AotAssemblies`
property when calling `RunApkTests.targets`