[Xamarin.Android.Build.Tasks] Apksigner invocation fails and produces build error when running against build tools 25 and lower
Fixes https://devdiv.visualstudio.com/DevDiv/_workitems?id=525752

Older `apksigner` versions did not have an `--in` parameter. Instead
they expected the last item to be passed to be the .apk to sign.

So to make this work on older `apksigner` versions we should rework
the Task to pass the apk last.