[tests] fix emulator launch when ANDROID_HOME is set
It appears the installation of Android Studio has exported `ANDROID_HOME`
on my system. This is problematic for our build, since it installs a
specific version of the Android SDK into ~/android-toolchain/sdk.

The fix is to set `ANDROID_HOME` for the `emulator` command using the
existing value for `AndroidSdkHome` from our build. This allows the emulator
to locate the proper system image in ~/android-toolchain/sdk/system-images.
I suspect a similar problem is happening on VSTS, in our monodroid build.

I also added the `-verbose` flag to the `emulator` command so we will have
more information if the emulator fails to launch during a build.