[tests] fix emulator launch when ANDROID_HOME is set (#1041)
It appears the installation of Android Studio has exported
`$ANDROID_HOME` on my system. This is problematic for our build,
as it installs a specific version of the Android SDK into
`$HOME/android-toolchain/sdk`.

The fix is to set `$ANDROID_HOME` for the `emulator` command using
the existing value for `$(AndroidSdkDirectory)` from our build. This
allows the emulator to locate the proper system image in
`$HOME/android-toolchain/sdk/system-images`.

I suspect a similar problem is happening on VSTS, in our
monodroid build.

Also added the `-verbose` flag to the `emulator` command so we will
have more information if the emulator fails to launch during a build.