[Xamarin.Android.Build.Tests] Make log searching more efficient. (#1034)
The current implementation of `LastBuildOutput` loads the entire
logfile into memory for searching. This causes OutOfMemory problems
on windows, especially now that we are running multiple tests at
the same time.

This commit switches the `LastBuildOutput` over to be an
`IEnumerable<string>`. The idea is that we will stream through
the buildlog one line at a time to search for the strings we
want to check. This might not be as quick, but it should be
more efficient.