[android-toolchain] update emulator to 26.1.4.0 (#1042)
Context: https://developer.android.com/studio/releases/emulator.html#26-1-3

Google has added support for `Hypervisor.framework` on Mac instead of
using HAXM. It appears all we need to do to enable this is use a newer
version of the emulator. Thanks to @JonDouglas for the info about this.

To check that `Hypervisor.framework` is enabled, run
`make. run-apk-tests V=1`, and you should see the following
output when the emulator starts:

    emulator: CPU Acceleration status: Hypervisor.Framework OS X Version 10.12

On my machine, I also had to uninstall HAXM, which you can either do by
running the silent uninstaller or the following command:

    sudo kextunload -b com.intel.kext.intelhaxm