Bump mono and work around duplicate tests issue
Need fix for https://bugzilla.xamarin.com/show_bug.cgi?id=60625

Unfortunately, the bump also exposed a problem with BCL tests. One test suite
(MonoTests.System.Security.Cryptography.CryptoConfigTest) is included in two
assemblies which our Nunit runner doesn't like as it assumes all the suites
have unique names and throws an exception when attempting to add the duplicate
suite to a dictionary indexed on name.

As a workaround (a new runner is in works), this patch renames the duplicate
suite by prepending assembly name as the prefix to its name.