[Xamarin.Android.Build.Tasks] New APT0000 msbuild errors are being reported from aapt output parsing
Fixes https://devdiv.visualstudio.com/DevDiv/_workitems?id=528572

The current RegEx was capturing output which was not and error
or warning. It turns out that aapt is not very consistent with
its error messaging. Sometimes they use `error`, sometimes then
use `Error` or `ERROR`. Or in the worse case, they don't flag it
as anything.

This commit adds a few more tests to AndroidRegExTests as well
as alters the logic in `Aapt` a bit as well. This logic will
only log as an error if we have an error `level` or we have
a `file` and `line`. This seems to be fairly consistent when
looking at the `aapt` source code.

We also added a few extra `Known` errors which the regex might not
pick. This is because they do not contain a `level` or a `file`
and `line` number.