[Xamarin.Android.Build.Tasks] Android design-time intellisense doesn't process global cache AARs
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=60878

The Managed Resource Parsing system was not including items downloaded
into the global resource cache. As a result intellisense for those
items did not work.

This commit fixes that by adding `_CreateAdditionalResourceCache` to the
`_ManagedUpdateAndroidResgen` target. This will ensure that those
items are included. Note that we still DO NOT download files during
the designtime build. So if the required items have not been downloaded
the intellisense will still not show those items. The user will
need to do a full build to ensure those items are in place.