[Xamarin.Android.Build.Tasks] Android design-time intellisense gets deleted by IncrementalClean
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=60880

For some weird reason, writing the designer.cs location to the
FileWrites file was not working. It was getting deleted still.

So lets add it to the Xamarin.Android.Windows.targets. This is what
we use to make sure files are NOT deleted by the IncrementalClean.
Again no idea why writing the data to the FileWrites file as NOT
working.

A unit test has also been updated to make sure the designtime
file is NOT removed on the next full build after the designtime
one.