[Xamarin.Android.Build.Tasks] IncrementalClean & R.d.cs (#1052)
Context: https://github.com/Microsoft/msbuild/issues/2408#issuecomment-321082997
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=60880

The `IncrementalClean` target will delete all files that MSBuild
doesn't "know about". MSBuild "knows about" a file if it is in
the file `$(IntermediateOutputPath)$(CleanFile)` and/or in the
`@(FileWrites)` item group (which is "serialized" into
`$(IntermediateOutputPath)$(CleanFile)`).

*However*, the contents of `$(CleanFile)` -- one file per line --
*must* contain *absolute paths*. If it doesn't contain absolute
paths, then it's not considered to be a "known file", and the
`IncrementalClean` target will *delete that file*.

Unfortunately, the `_ManagedUpdateAndroidResgen` target was writing
a *relative* path to `Resource.Designer.cs` into `$(CleanFile)`. This
resulted in the Visual Studio DesignTimeBuild (and related builds)
*deleting* `Resource.Designer.cs`, preventing code completion on the
generated `Resource` type from behaving as expected.

Update the `_ManagedUpdateAndroidResgen` target so that the absolute
path to `Resource.Designer.cs` is written into `$(CleanFile)`. This
prevents `Resource.Designer.cs` from being inadvertently deleted by
the `IncrementalClean` target.