[Xamarin.ProjectTools] Fix an IO error when reading process.log (#1051)
If we hit a native crash in both of the attempts to run an MSBuild
test we get an IO exception when trying to append `process.log` to
the build output:

	System.IO.FileNotFoundException: Could not find file process.log

This is because `process.log` was moved to `process.log.bak` when we
get a native crash.

What we want is to ignore any native crashes on the first run of the
test. If we still get a native crash on the next run we should log it.

So on the first crash we move the file over to the `process.log.bak`.
But we only do this on the first crash (attempt == 0). Otherwise we
end up moving the file again and we hit an error when trying to
append the `process.log` to the build output.