[Xamarin.Android.Build.Tasks] APT0000 errors are for *errors* (#1047)
Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems?id=528572

The current Regex was capturing output which was not an error or a
warning. It turns out that `aapt` is not very consistent with its
error messaging: sometimes it uses `error`, other times it uses
`Error` or `ERROR`.

Or in the worse case, they don't flag it as anything.

This commit adds a few more tests to `AndroidRegExTests` as well as
alters the logic in the `<Aapt/>` task a bit as well. This logic will
only log an error if:

 1. We have an error `level` group, or
 2. we have a `file` name group and `line` number group.

This seems to be fairly consistent when looking at the `aapt` source
code.

We also update the tests to add a few extra `Known` errors which the
regex might not pick. This is because they do not contain a `level`
or a `file` and `line` number.