[Xamarin.Android.Build.Tasks] Add LogCustomBuildEvent method to AsyncTask
There is a possilility that we might be locking up the UI thread
when logging CustomBuildEvents.. This happens in the
`InstallPackageAssemblies` task. Currently we only use

	BuildEngine.LogCustomEvent

Which may well cause problems when called from a non UI thread.

So in keeping with the other Log methods available we should add
a new LogCustomBuildEvent. This will queue and marshall the data
to the UI Thread just like the other methods do.