[Xamarin.Android.Build.Tasks] Fix IncrementalClean on msbuild. (#1063)
Context: https://bugzilla.xamarin.com/show_bug.cgi?id=60880

Commit 552ad5d7 added code to make sure the `designtime` build
files are NOT deleted by the `IncrementalClean` target.
While it fixed the issue for `xbuild` it did NOT fix it for
`msbuild`. As a result the file was still deleted.

So we need to add the file to the target in the
Xamarin.Android.Windows.targets to ensure that it is NOT
removed (just like the mdb files).