[Xamarin.Android.Build.Tasks] Add parsing of R.txt (#1058)
In order to speed up the processing of resources we should
make use of the R.txt files which are generally included with
.aar files and our components.

One advantage with this is we get the actual resource id which
is used by android. But it also means that when we do find an
R.txt file we can skip the processing of ALL resources in that
underlying `res` folder. Note this will only apply to the
ManagedResourceParser.