[Xamarin.ProjectTools] Only add the process.log if it exists. (#1070)
This commit adds an additional check to only process the
process.log if the file exists. We might be in a situation
where nothing is written to StdOut or StdErr as a result we
might not get a log file. So lets not crash the test if it
does not exist.