[Xamarin.Android.Build.Tests] Fix detection of Emulators.
If the `adb` command times out it will probably return
an empty string. In that case we should NOT run the
tests which need devices.

So instead of checking for an error, lets try to parse
the output to an int. We should be getting a valid
integer value back for the sdk version. If we do not then
we should not run those tests.